created clawTuner.java and modified Claw.java to add setClawPos method
I made a simple op mode which allows you to move the claw with a joystick y axis and then displays the current servo value in the telemetry. In order to make this work, I added a new method to the claw class to allow setting the servo position manually. In order to make this work, we still need to update the hardware map on the driver hub.